Allow the record data to vary in chunk size
This code replaces the logic that buffers the audio record samples as
the existing code would only allow streaming buffers of 4096 frames at
a time causing 85.5ms of latency. This new code replaces the fixed size
ringbuffer with a variable size allowing any size write or read up to
the maximum buffer size.

Signed-off-by: Geoffrey McRae <geoff@hostfission.com>